feat: create custom hook, rendered toaster in appshell
