+++ apps/gitness/src/pages-v2/repo/repo-list.tsx
@@ -1,19 +1,17 @@
-import { useCallback, useEffect, useMemo, useState } from 'react'
-import { Link, useParams } from 'react-router-dom'
+import { useEffect } from 'react'
+import { useParams } from 'react-router-dom'
 
-import { RepoRepositoryOutput, useDeleteRepositoryMutation, useListReposQuery } from '@harnessio/code-service-client'
+import { useDeleteRepositoryMutation, useListReposQuery } from '@harnessio/code-service-client'
 import { ToastAction, useToast } from '@harnessio/ui/components'
 import { RepositoryType, SandboxRepoListPage } from '@harnessio/ui/views'
 
-import { Toaster } from '../../components-v2/toaster'
 import { useRoutes } from '../../framework/context/NavigationContext'
 import { useGetSpaceURLParam } from '../../framework/hooks/useGetSpaceParam'
 import { useQueryState } from '../../framework/hooks/useQueryState'
-import useSpaceSSE from '../../framework/hooks/useSpaceSSE'
 import usePaginationQueryStateWithStore from '../../hooks/use-pagination-query-state-with-store'
 import { useTranslationStore } from '../../i18n/stores/i18n-store'
 import { PathParams } from '../../RouteDefinitions'
-import { PageResponseHeader, SSEEvent } from '../../types'
+import { PageResponseHeader } from '../../types'
 import { useRepoStore } from './stores/repo-list-store'
 import { transformRepoList } from './transform-utils/repo-list-transform'
 
@@ -27,12 +25,10 @@ export default function ReposListPage() {
     setPage,
     importRepoIdentifier,
     setImportRepoIdentifier,
-    updateRepository,
     importToastId,
     setImportToastId
   } = useRepoStore()
-  // const [importToastId, setImportToastId] = useState<string | null>(null)
-  const { toast, dismiss, update } = useToast()
+  const { toast, dismiss } = useToast()
 
   const [query, setQuery] = useQueryState('query')
   const { queryPage } = usePaginationQueryStateWithStore({ page, setPage })
@@ -102,43 +98,10 @@ export default function ReposListPage() {
           </ToastAction>
         )
       })
-      console.log('importRepoIdentifier', importRepoIdentifier, importToastId, id)
 
       setImportToastId(id)
     }
   }, [importRepoIdentifier, setImportRepoIdentifier])
-
-  const onEvent = useCallback(
-    (eventData: RepoRepositoryOutput) => {
-      update({
-        id: importToastId ?? '',
-        title: 'Repository imported',
-        description: importRepoIdentifier,
-        duration: 5000,
-        action: (
-          <Link to={routes.toRepoSummary({ spaceId, repoId: importRepoIdentifier ?? '' })}>
-            <ToastAction altText="View repository">View</ToastAction>
-          </Link>
-        ),
-        variant: 'success'
-      })
-      const transformedRepo = transformRepoList([eventData])
-      updateRepository(transformedRepo[0])
-      // addRepository(transformedRepo[0])
-      setImportRepoIdentifier(null)
-      setImportToastId(null)
-    },
-    [importToastId]
-  )
-
-  const events = useMemo(() => [SSEEvent.REPO_IMPORTED], [])
-
-  useSpaceSSE({
-    space: spaceURL,
-    events,
-    onEvent,
-    shouldRun: true
-  })
 
   return (
     <>
@@ -154,7 +117,6 @@ export default function ReposListPage() {
         toCreateRepo={() => routes.toCreateRepo({ spaceId })}
         toImportRepo={() => routes.toImportRepo({ spaceId })}
       />
-      <Toaster />
     </>
   )
 }
