+++ packages/ui/locales/es/views.json
@@ -171,7 +171,6 @@
     "updateWebhook": "Update webhook",
     "createWebhook": "Create webhook",
     "tagger": "Tagger",
-
     "renameBranch": "Rename Branch",
     "add": "Add",
     "users": "Users",
