+++ packages/ui/locales/fr/views.json
@@ -171,7 +171,6 @@
     "updateWebhook": "Update webhook",
     "createWebhook": "Create webhook",
     "tagger": "Tagger",
-
     "renameBranch": "Renommer la branche",
     "add": "Ajouter",
     "users": "Utilisateurs",
